Olympics dataset
The following code loads the Olympics dataset (olympics.csv), which was derived from the
Wikipedia entry on All-Time Olympics Games Medals., and does some basic data cleaning.
The columns are organized as# of Summer games, Summer medals, # of Winter games, Winter
medals, total # number of games, and total # of medals. Use this dataset to answer the questions
below.